requirements.txt
dnspython
requests
colorama
